requirements.txt
# Core libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# ML libraries
xgboost==1.7.6
tensorflow==2.13.0
keras==2.13.1

# Signal processing
scipy==1.11.1
pywavelets==1.4.1

# Dashboard
streamlit==1.25.0

# Utilities
joblib==1.3.1
pyyaml==6.0.1
tqdm==4.65.0
requests==2.31.0